update logout api

+++ src/actions/auth.Action.js
@@ -59,9 +59,25 @@ export const isUserLoggedIn = () => {
 
 export const signout = () => {
    return async (dispatch) => {
-      window.localStorage.clear();
       dispatch({
          type: authConstants.LOGOUT_REQUEST,
       });
+
+      const res = await axios.post(`admin/signout`);
+
+      if (res.status === 200) {
+         window.localStorage.clear();
+
+         dispatch({
+            type: authConstants.LOGOUT_SUCCESS,
+         });
+      } else {
+         dispatch({
+            type: authConstants.LOGOUT_FAILURE,
+            payload: {
+               error: "Failed to logout",
+            },
+         });
+      }
    };
 };
